backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.5.0
python-socketio>=5.11.0
requests>=2.31.0
APScheduler>=3.10.4
google-generativeai>=0.3.0
pillow>=10.0.0
python-dotenv>=1.0.0
yt-dlp>=2024.1.0
gunicorn>=21.0.0
eventlet>=0.35.0
joblib>=1.3.0
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
